terraform: move module to its own file
